12.0.8.0

+++ com/baidu/tieba/sharesdk/a/d.java
@@ -6,7 +6,7 @@ import android.text.TextUtils;
 import com.baidu.adp.lib.util.l;
 import com.baidu.tieba.R;
 import com.baidu.tieba.sharesdk.bean.ShareEntity;
-/* loaded from: classes24.dex */
+/* loaded from: classes23.dex */
 public class d extends a {
     public d(Context context) {
         super(context);
@@ -16,7 +16,7 @@ public class d extends a {
     public void a(ShareEntity shareEntity, com.baidu.tieba.sharesdk.b.b bVar) {
         String str;
         if (shareEntity == null || TextUtils.isEmpty(shareEntity.getContent())) {
-            l.showToast(getAppContext(), (int) R.string.share_empty_tip);
+            l.showToast(getAppContext(), R.string.share_empty_tip);
             if (bVar != null) {
                 bVar.dn(0, 2);
                 return;
